+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add Datadog listener for metrics + errors publishing
 - Add Datadog example dashboard template
 - Update Readme to show Dd instrumentation usage
+- Align the directory namespace convention with gem name (waterdrop => WaterDrop)
 
 ## 2.1.0 (2022-01-03)
 - Ruby 3.1 support
