+++ lib/waterdrop/contracts.rb
