+++ lib/waterdrop/contracts/config.rb
