+++ lib/waterdrop/contracts/message.rb
