+++ lib/waterdrop/errors.rb
