+++ lib/waterdrop/instrumentation.rb
