+++ lib/waterdrop/instrumentation/callbacks/error.rb
