+++ lib/waterdrop/instrumentation/callbacks/statistics.rb
