+++ lib/waterdrop/instrumentation/callbacks/statistics_decorator.rb
