+++ lib/waterdrop/instrumentation/callbacks_manager.rb
