+++ lib/waterdrop/instrumentation/monitor.rb
