+++ lib/waterdrop/instrumentation/stdout_listener.rb
