+++ lib/waterdrop/patches/rdkafka/bindings.rb
