+++ lib/waterdrop/patches/rdkafka/producer.rb
