+++ lib/waterdrop/producer/buffer.rb
