+++ lib/waterdrop/producer/dummy_client.rb
