+++ lib/waterdrop/producer/status.rb
