+++ lib/waterdrop/version.rb
